Try fixing verification again
