+++ packages/login/next.config.js
@@ -1,0 +1,22 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  webpack(config, { isServer }) {
+    if (!isServer) {
+      config.resolve.fallback = {
+        ...config.resolve.fallback,
+
+        fs: false, //"@pcd/gpc" relies on fs but we use some fns client side
+      };
+      config.resolve.alias = {
+        constants: require.resolve(
+          "rollup-plugin-node-polyfills/polyfills/constants"
+        ),
+        process: "process/browser"
+      };
+    }
+
+    return config;
+  },
+};
+
+module.exports = nextConfig;
