+++ packages/login/package.json
@@ -32,6 +32,7 @@
 		"eslint-config-next": "14.2.15",
 		"patch-package": "^8.0.0",
 		"postcss": "^8",
+		"rollup-plugin-node-polyfills": "^0.2.1",
 		"tailwindcss": "^3.4.1",
 		"typescript": "^5"
 	}
